Update weirdAAL.py
Fixed: ImportError: cannot import name 'Iterable' from 'collections'
for python3.11 version

The error occurs because the `weirdAAL.py` script attempts to import the Iterable class from the collections module, which does not include this class in Python 3.11. In Python 3.11, the Iterable class is included in the collections.abc module instead.